Have mega refactoring working
