+++ src/webparts/drilldown7/CoreFPS/BuildExportProps.ts
@@ -27,11 +27,8 @@
   *                                                                                                                                                                              
   */
 
- import { getHelpfullError } from '@mikezimm/npmfunctions/dist/Services/Logging/ErrorHandler';
- import { getHelpfullErrorV2, saveThisLogItem } from  '@mikezimm/npmfunctions/dist/Services/Logging/ErrorHandler';
-
  import { createExportObject, } from '@mikezimm/npmfunctions/dist/Services/PropPane/ExportFunctions';
- import { IMinExportObject, } from '@mikezimm/npmfunctions/dist/Services/PropPane/ExportFunctions';
+
 
   /***
   *    d888888b .88b  d88. d8888b.  .d88b.  d8888b. d888888b       .o88b.  .d88b.  .88b  d88. d8888b.  .d88b.  d8b   db d88888b d8b   db d888888b 
@@ -45,37 +42,17 @@
   */
 
 
-import { IDrilldown7WebPartProps } from '../IDrilldown7WebPartProps';
-import { exportIgnoreProps, importBlockProps, } from '../IDrilldown7WebPartProps';
-
-import { changeVisitor, changeExpando, changeBanner, changefpsOptions1, changefpsOptions2, exportIgnorePropsFPS, importBlockPropsFPS,  } from '@mikezimm/npmfunctions/dist/WebPartInterfaces/ImportProps';
-
-import { changeBannerBasics, changeBannerNav, changeBannerTheme, changeBannerOther,  } from '@mikezimm/npmfunctions/dist/WebPartInterfaces/ImportProps';
-
 // import { ILoadPerformanceALVFM, IPerformanceOp } from './components/Performance/IFpsPageInfoWebPartProps';
 
+import { IDrilldown7WebPartProps, exportIgnoreProps, } from '../IDrilldown7WebPartProps';
 
+import { changeCustomHelp, changeExpando, changeBanner, changePageStyle, changefpsOptions2,  } from '../fpsReferences';
 
+import { changeBannerBasics, changeBannerNav, changeBannerTheme, changeBannerUtility,  } from '../fpsReferences';
 
-/***
- *    .d8888.  .d8b.  .88b  d88. d8888b. db      d88888b       .d88b.  d8b   db db      db    db                                                                            
- *    88'  YP d8' `8b 88'YbdP`88 88  `8D 88      88'          .8P  Y8. 888o  88 88      `8b  d8'                                                                            
- *    `8bo.   88ooo88 88  88  88 88oodD' 88      88ooooo      88    88 88V8o 88 88       `8bd8'                                                                             
- *      `Y8b. 88~~~88 88  88  88 88~~~   88      88~~~~~      88    88 88 V8o88 88         88                                                                               
- *    db   8D 88   88 88  88  88 88      88booo. 88.          `8b  d8' 88  V888 88booo.    88                                                                               
- *    `8888Y' YP   YP YP  YP  YP 88      Y88888P Y88888P       `Y88P'  VP   V8P Y88888P    YP                                                                               
- *                                                                                                                                                                          
- *                                                                                                                                                                          
- *     .o88b.  .d88b.  d8888b. db    db      d888888b  .d88b.       db       .d88b.   .o88b.  .d8b.  db           d8888b. d8888b.  .d88b.     d88b d88888b  .o88b. d888888b 
- *    d8P  Y8 .8P  Y8. 88  `8D `8b  d8'      `~~88~~' .8P  Y8.      88      .8P  Y8. d8P  Y8 d8' `8b 88           88  `8D 88  `8D .8P  Y8.    `8P' 88'     d8P  Y8 `~~88~~' 
- *    8P      88    88 88oodD'  `8bd8'          88    88    88      88      88    88 8P      88ooo88 88           88oodD' 88oobY' 88    88     88  88ooooo 8P         88    
- *    8b      88    88 88~~~      88            88    88    88      88      88    88 8b      88~~~88 88           88~~~   88`8b   88    88     88  88~~~~~ 8b         88    
- *    Y8b  d8 `8b  d8' 88         88            88    `8b  d8'      88booo. `8b  d8' Y8b  d8 88   88 88booo.      88      88 `88. `8b  d8' db. 88  88.     Y8b  d8    88    
- *     `Y88P'  `Y88P'  88         YP            YP     `Y88P'       Y88888P  `Y88P'   `Y88P' YP   YP Y88888P      88      88   YD  `Y88P'  Y8888P  Y88888P  `Y88P'    YP    
- *                                                                                                                                                                          
- *                                                                                                                                                                          
- */
+import { changePinMe,  } from '../fpsReferences';
 
+import { importBlockProps } from '../IDrilldown7WebPartProps';
 
 /***
  *    d88888b db    db d8888b.  .d88b.  d8888b. d888888b      d8888b. d8888b.  .d88b.  d8888b. .d8888. 
@@ -93,8 +70,6 @@ import { changeBannerBasics, changeBannerNav, changeBannerTheme, changeBannerOth
  * @returns exportObject
  */
 
-import {  changeWebPartStyles, changeRelated1 } from '../IDrilldown7WebPartProps';
-
  export function buildExportProps( wpProps : IDrilldown7WebPartProps, wpInstanceID: string, currentWeb: string, ) {
     let exportStructure :any = {};
     // let wpInstanceIDSplit = wpInstanceID.split('|');
@@ -110,19 +85,20 @@ import {  changeWebPartStyles, changeRelated1 } from '../IDrilldown7WebPartProps
     // exportStructure.changeRelated2 = changeRelated2;
     // exportStructure.pageLinks = changePageLinks;
 
-    exportStructure.changeWebPartStyles = changeWebPartStyles;
+    exportStructure.Visitor = changeCustomHelp;
 
-    // exportStructure.Script = changeScript;
+    exportStructure.BannerBasics = changeBannerBasics;
+    exportStructure.BannerNav = changeBannerNav;
 
-    exportStructure.Visitor = changeVisitor;
+    exportStructure.BannerTheme = changeBannerTheme;
+    exportStructure.BannerOther = changeBannerUtility;
 
-    exportStructure.Banner = changeBanner;
-
-    exportStructure.fpsOptions1 = changefpsOptions1;
+    exportStructure.fpsOptions1 = changePageStyle;
     
     exportStructure.Expando = changeExpando;
 
     exportStructure.fpsOptions2 = changefpsOptions2;
+
 
     let exportObject = createExportObject( exportStructure, wpProps, exportIgnoreProps, false );
 
@@ -137,17 +113,18 @@ import {  changeWebPartStyles, changeRelated1 } from '../IDrilldown7WebPartProps
     exportStructure.wpInstanceID = wpInstanceID;
     exportStructure.currentWeb = currentWeb;
 
-    exportStructure.Visitor = changeVisitor;
+    exportStructure.Visitor = changeCustomHelp;
 
     exportStructure.BannerBasics = changeBannerBasics;
     exportStructure.BannerNav = changeBannerNav;
 
     exportStructure.BannerTheme = changeBannerTheme;
-    exportStructure.BannerOther = changeBannerOther;
+    exportStructure.BannerOther = changeBannerUtility;
 
-    // exportStructure.Expando = changeExpando;
+    exportStructure.fpsOptions1 = changePageStyle;
 
-    exportStructure.fpsOptions1 = changefpsOptions1;
+    exportStructure.Expando = changeExpando;
+
     exportStructure.fpsOptions2 = changefpsOptions2;
 
     let exportObject = createExportObject( exportStructure, wpProps, exportIgnoreProps, false );
