+++ src/webparts/drilldown7/components/CssCharts/cssChartFunctions.ts
@@ -1,8 +1,7 @@
 import { toAbsoluteUrl } from '@pnp/sp';
-import { ICSSChartSettings, ICSSChartData, ICSSChartTypes, CSSChartTypes, ISeriesSort } from '../IReUsableInterfaces';
+import { ICSSChartSettings, ICSSChartData, } from '../Drill/IDrillProps';
 
-import { IRefinerLayer, IRefiners, IItemRefiners, IRefinerStats, RefineRuleValues,
-    IRefinerRules, IRefinerStatType, RefinerStatTypes, IRefinerStat } from '@mikezimm/npmfunctions/dist/Refiners/IRefiners';
+import { IRefinerLayer, IRefinerStat, ICSSChartTypes } from '../../fpsReferences';
 
 export function buildCountChartsObject( title: string, callBackID: string, refinerObj: IRefinerLayer , chartTypes: ICSSChartTypes[] ) {
     let resultSummaryObject = null;
