+++ src/webparts/drilldown7/components/Drill/drillFunctions.tsx
@@ -13,7 +13,7 @@ import { IContentsListInfo, IMyListInfo, IServiceLog, IContentsLists } from '../
 
 import { IDrillItemInfo } from '@mikezimm/npmfunctions/dist/WebPartInterfaces/DrillDown/IDrillItem';
 
-import { IDrillList, pivCats } from  './drillComponent';
+import { IDrillList, } from  './IDrillProps';
 
 
 import { changes, IMyFieldTypes } from '../../../../services/listServices/columnTypes'; //Import view arrays for Time list
