+++ src/webparts/drilldown7/components/Drill/drillListView.tsx
@@ -2,8 +2,7 @@
 import * as React from 'react';
 import { Icon  } from 'office-ui-fabric-react/lib/Icon';
 
-//import { IDrillItemInfo } from '@mikezimm/npmfunctions/dist/WebPartInterfaces/DrillDown/IDrillItem';
-import { IDrillItemInfo } from '@mikezimm/npmfunctions/dist/WebPartInterfaces/DrillDown/IDrillItem';
+import { IDrillItemInfo } from '../../fpsReferences';
 
 import { buildPropsHoverCard } from '../../../../services/hoverCardService';
